@@ -1,22 +1,33 @@
 {
-	"name": "dface/sql-placeholders",
-	"description": "PHP classes to parse/build sql-expressions extended with special syntax",
-	"keywords": ["SQL", "placeholder"],
-	"type": "library",
-	"license": "MIT",
-	"authors": [{
-		"name": "Denis Ponomarev",
-		"email": "ponomarev@gmail.com"
-	}
-	],
-	"require": {
-		"php": ">=5.6.0",
-		"ext-ctype": "*"
-	},
-	"require-dev": {
-		"phpunit/phpunit": "3.7.*"
-	},
-	"autoload": {
-		"psr-0": {"dface": "src/"}
-	}
+  "name": "dface/sql-placeholders",
+  "description": "Parse and build sql-expressions extended with special syntax",
+  "keywords": [
+    "SQL",
+    "placeholder"
+  ],
+  "type": "library",
+  "license": "MIT",
+  "authors": [
+    {
+      "name": "Denis Ponomarev",
+      "email": "ponomarev@gmail.com"
+    }
+  ],
+  "require": {
+    "php": "^7.4",
+    "ext-ctype": "*"
+  },
+  "require-dev": {
+    "phpunit/phpunit": "^9.2"
+  },
+  "autoload": {
+    "psr-0": {
+      "dface\\sql\\placeholders\\": "src/"
+    }
+  },
+  "autoload-dev": {
+    "psr-0": {
+      "dface\\sql\\placeholders\\": "tests/"
+    }
+  }
 }
